Update the Move Windows feature
"Move Windows" Feature V1 (First Code Snippet)
The "Move Windows" Feature V1 allows users to move windows (frames) in the Hyggshi OS Dev interface on Roblox Studio by dragging and dropping using the mouse or touch. This feature allows windows to be repositioned intuitively and easily, providing a smoother user experience. Here is a detailed description of how this feature works:

Detailed Description:
1 Structure and Initialization:

This code will be associated with a Frame (window) that you want to allow the user to move.

This window must be assigned as the parent object of the code, i.e. script.Parent.Parent needs to point to a Frame object that you want to move.

2 Required Variables:

dragging: This variable keeps track of the state when the user is dragging the window.

dragInput: This variable stores the input object when the user performs a drag operation.
dragStart: The initial position when the user starts dragging the window.

startPos: The initial position of the window before moving.

3 Update function:

The update function calculates the change between the starting position of the drag operation and the current position of the cursor (or finger) to determine the new position of the window.

This function will change the Position property of the Frame so that the window moves smoothly.

4 Connect the InputBegan event:

When the user starts clicking (or touching), the code will set the dragging variable to true to indicate that the user is dragging.

At this time, the code also records the initial position of the cursor/touch and the current position of the window.

5 Connect the InputChanged event:

When the user moves the mouse or touch while dragging, the InputChanged event will record the dragInput object (the mouse/touch is moving).

6 Updating and Moving the Window:

Finally, in the InputChanged event from the UserInputService, the code checks to see if the user is dragging (the dragging variable is true) and if there is a change in the cursor (or touch) position. If so, the update function is called, and the window is moved.

7 Ending the Drag:

When the user finishes dragging (releases the mouse or finger), the code stops tracking the dragging state by setting dragging to false.